Adding Logo component and some refactor
