+++ src/components/Logo/styles.module.scss
@@ -1,0 +1,7 @@
+@import "~$styles/_variables";
+
+.logo {
+  height: $logoHeight;
+  margin: $spacing-m;
+  width: $logoWidth;
+}
