+++ src/components/NavBar/styles.module.scss
@@ -21,12 +21,6 @@
   }
 }
 
-.logo {
-  height: $navBarLogoHeight;
-  margin: $spacing-m;
-  width: $navBarLogoWidth;
-}
-
 .link {
   color: $white;
   text-decoration: none;
