+++ src/pages/CreateVolunteer/container.tsx
@@ -3,6 +3,7 @@ import { FunctionComponent } from "react";
 import { CreateVolunteer } from "./component";
 import { IUseCreateVolunteerVariables, useCreateVolunteer } from "$hooks";
 import { useHistory } from "react-router-dom";
+import { RoutesBuilder } from "$models";
 
 export const CreateVolunteerContainer: FunctionComponent = () => {
   const { createVolunteer } = useCreateVolunteer();
@@ -10,7 +11,7 @@ export const CreateVolunteerContainer: FunctionComponent = () => {
 
   const onSubmit = async (variables: IUseCreateVolunteerVariables) => {
     await createVolunteer(variables);
-    history.push("/");
+    history.push(RoutesBuilder.volunteers.home());
   };
   return (
     <CreateVolunteer
