+++ src/styles/_variables.scss
@@ -14,8 +14,8 @@ $fieldWeight: 100%;
 $headerHeight: 150px;
 $buttonHeight: 77px;
 $navBarHeight: 60px;
-$navBarLogoHeight: 60px;
-$navBarLogoWidth: 152px;
+$logoHeight: 60px;
+$logoWidth: 152px;
 $pageMargin: 1.5vw;
 $tableMargin: 2px;
 $borderTable: 1px solid;
